% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calls_from_vcf.R
\name{add_calls_from_vcf}
\alias{add_calls_from_vcf}
\title{add minor allele counts from a VCF to SummarizedExperiment with common samples}
\usage{
add_calls_from_vcf(se, vcfpath, variantRanges)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{vcfpath}{path to VCF with genotypes for samples}

\item{variantRanges}{range to filter variants before adding to SummarizedExperiment}
}
\value{
an updated SummarizedExperiment with metadata elements snpaddrs and snpchrs and colData
augmented with genotype calls.
}
\description{
add minor allele counts from a VCF to SummarizedExperiment with common samples
}
\note{
If a metadata element nonCallVars is present, the function will exit with error.
}
\examples{
vp = system.file("vcf", "chr19_50k.vcf.gz", package="biocXqtl")
data("mageSE_19", package="biocXqtl")
nn = add_calls_from_vcf(mageSE_19, vp, GRanges("19:1-150000"))
}
