% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loc_trans.R
\name{add_loc_trans}
\alias{add_loc_trans}
\title{utility to add annotation to trans results}
\usage{
add_loc_trans(xexp, trout)
}
\arguments{
\item{xexp}{XqtlExperiment instance}

\item{trout}{output of trans.map_trans}
}
\value{
data.frame instance with variant and phenotype feature annotation as
new columns.  Note that there may be some excess conversion to character class.
}
\description{
utility to add annotation to trans results
}
\note{
This started to fail on left_join because a pandas DataFrame will not autoconvert.
Probably need pandas 2.3.3
}
