% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_Zs.R
\name{add_scores}
\alias{add_scores}
\title{internal function to combine new xQTL test results with existing XqtlExperiment instance}
\usage{
add_scores(xse, mat, check.cols = TRUE)
}
\arguments{
\item{xse}{XqtlExperiment instance}

\item{mat}{a matrix of test results conformant with the XqtlExperiment: rownames agree,}

\item{check.cols}{logical(1) will message if redundant variants are found when new scores
are bound in
columns will be added to the mcols(rowRanges(xse))}
}
\description{
internal function to combine new xQTL test results with existing XqtlExperiment instance
}
