% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_Zs.R
\name{bind_Zs}
\alias{bind_Zs}
\title{compute Z-statistics for xQTL association and bind them to the input x-ome+genome XqtlExperiment}
\usage{
bind_Zs(xse, omit.hi.maf = FALSE, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{xse}{XqtlExperiment instance}

\item{omit.hi.maf}{logical(1), passed to zs4manyYs, defaults to FALSE}

\item{BPPARAM}{instance of BiocParallelParam}
}
\description{
compute Z-statistics for xQTL association and bind them to the input x-ome+genome XqtlExperiment
}
\examples{
data(geuv19xse)
lk = geuv19xse[1:20,]
mafs = maf(lk) # only snps here
mins = apply(data.matrix(mcols(getCalls(lk))), 1, min, na.rm=TRUE) # some -1 values
#filterCalls = function(xse, keep) { slot(xse, "calls") = slot(xse, "calls")[keep,]; xse }
lk = filterCalls(lk, which(mafs>.25 & mins > -1))
lk = bind_Zs(lk)
head(rowRanges(lk)[,7])
plpp2tx = as.numeric(assay(lk["ENST00000434325",]))
hiz = as.numeric(data.matrix(mcols(getCalls(lk))["snp_19_5694231",]))
plot(plpp2tx~jitter(hiz), xlab="snp at 19:5694231", ylab="counts for a transcript of PLPP2")
summary(lm(plpp2tx~hiz))
}
