% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getzs.R
\name{bind_Zs_old}
\alias{bind_Zs_old}
\title{bind a matrix of Z statistics created with getzs to the rowRanges of a SummarizedExperiment}
\usage{
bind_Zs_old(se)
}
\arguments{
\item{se}{SummarizedExperiment assumed to have molecular phenotype data in assay.  A metadata
component (list element) named nonCallVars will be checked and associated colData elements
will be used as covariates in models for effect of dosage of minor allele}
}
\description{
bind a matrix of Z statistics created with getzs to the rowRanges of a SummarizedExperiment
}
\note{
All variables in colData(se) will be used as covariates in association tests.  Use
`colData(se) <- NULL` to eliminate covariate adjustment.
}
\examples{
data(geuv19xse)
lk = geuv19xse[1:20,]
mafs = maf(lk) # only snps here
mins = apply(data.matrix(mcols(getCalls(lk))), 1, min, na.rm=TRUE) # some -1 values
#filterCalls = function(xse, keep) { slot(xse, "calls") = slot(xse, "calls")[keep,]; xse }
lk = filterCalls(lk, which(mafs>.25 & mins > -1))
lk = bind_Zs(lk)
head(rowRanges(lk)[,7])
plpp2tx = as.numeric(assay(lk["ENST00000434325",]))
hiz = as.numeric(data.matrix(mcols(getCalls(lk))["snp_19_5694231",]))
plot(plpp2tx~jitter(hiz), xlab="snp at 19:5694231", ylab="counts for a transcript of PLPP2")
summary(lm(plpp2tx~hiz))
}
