% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocXqtl-package.R
\docType{package}
\name{biocXqtl-package}
\alias{biocXqtl}
\alias{biocXqtl-package}
\title{biocXqtl: modular system for creating molecular QTL statistics}
\description{
This is a modular system for creating molecular QTL statistics. Given a RangedSummarizedExperiment with assay elements comprised of molecular responses like RNA-seq counts, and genotype calls in colData, results of association tests will be bound into the rowData of the input RangedSummarizedExperiment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vjcitn/biocXqtl}
  \item Report bugs at \url{https://github.com/vjcitn/biocXqtl/issues}
}

}
\author{
\strong{Maintainer}: Vince Carey \email{stvjc@channing.harvard.edu} (\href{https://orcid.org/0000-0003-4046-0063}{ORCID})

}
\keyword{internal}
