% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{countGenotypes}
\alias{countGenotypes}
\title{Count genotypes across samples for each variant}
\usage{
countGenotypes(gt)
}
\arguments{
\item{gt}{matrix of genotype calls (variants x samples) with form "0|0" or "0/0",
as produced by VariantAnnotion::geno applied to a VCF object}
}
\value{
DataFrame with counts for each genotype category
}
\description{
Count genotypes across samples for each variant
}
