% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{countGenotypesVectorized}
\alias{countGenotypesVectorized}
\title{More efficient vectorized version for large datasets}
\usage{
countGenotypesVectorized(gt)
}
\arguments{
\item{gt}{matrix of genotype calls (variants x samples) with form "0|0" or "0/0"
as produced by VariantAnnotion::geno applied to a VCF object}
}
\value{
DataFrame with counts for each genotype category
}
\description{
More efficient vectorized version for large datasets
}
