% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_vcf.R
\name{demo_vcf}
\alias{demo_vcf}
\title{provide path to a demonstration VCF file}
\usage{
demo_vcf()
}
\description{
provide path to a demonstration VCF file
}
\note{
derived from output of
`aws s3 cp --no-sign-request s3://1000genomes/release/20130502/ALL.chr19.phase3_shapeit2_mvncall_integrated_v5a.20130502.genotypes.vcf.gz .` 26 Dec 2025.

Variants from positions 1-5000000 were filtered by restricting to samples in
`mageSE_19` and variants with at least 10 individuals in each genotype class (00, 01, 11).
}
\examples{
v = demo_vcf()
VariantAnnotation::scanVcfHeader(v)
}
