% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tensorqtl.R
\name{example_tensorQTL}
\alias{example_tensorQTL}
\title{produce arrow dataset with tensorQTL cis example output}
\usage{
example_tensorQTL()
}
\description{
produce arrow dataset with tensorQTL cis example output
}
\examples{
if (requireNamespace("arrow")) {
example_tensorQTL()
}
}
