% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{filterCalls}
\alias{filterCalls}
\title{filter the call component of an XqtlExperiment}
\usage{
filterCalls(xse, keep)
}
\arguments{
\item{xse}{XqtlExperiment instance}

\item{keep}{vector of indices or logical selections}
}
\description{
filter the call component of an XqtlExperiment
}
