% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{filterVcfByGenotypeCounts}
\alias{filterVcfByGenotypeCounts}
\title{Filter VCF to variants with minimum counts in each genotype group}
\usage{
filterVcfByGenotypeCounts(
  vcf_file,
  samples = NULL,
  min_per_genotype = 1,
  genome = "hg38",
  regions = NULL,
  verbose = TRUE
)
}
\arguments{
\item{vcf_file}{Path to tabix-indexed VCF file (.vcf.gz with .tbi index)}

\item{samples}{Character vector of sample IDs to retain}

\item{min_per_genotype}{Minimum number of individuals required in each 
genotype group (homozygous reference, heterozygous, homozygous alternate)}

\item{genome}{Genome identifier (e.g., "hg19", "hg38")}

\item{regions}{Obligatory GRanges object to restrict to specific genomic regions}

\item{verbose}{Logical, print progress messages}
}
\value{
A filtered VCF object
}
\description{
Filter VCF to variants with minimum counts in each genotype group
}
\examples{
filtvcf <- filterVcfByGenotypeCounts(
  vcf_file = demo_vcf(),
  samples = c("HG00096", "HG00100", "HG00105", "HG00108", "HG00110", "HG00113", 
     "HG00114", "HG00121", "HG00122", "HG00127", "HG00130", "HG00132", 
     "HG00141", "HG00142", "HG00146", "HG00148", "HG00151", "HG00160", 
     "HG00177", "HG00179"),
  min_per_genotype = 2,
  regions = GenomicRanges::GRanges("19:1-50000000"),
  genome = "hg38"
)
filtvcf
}
