% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{filterVcfMultipleCriteria}
\alias{filterVcfMultipleCriteria}
\title{Filter VCF with multiple criteria}
\usage{
filterVcfMultipleCriteria(
  vcf,
  min_per_genotype = NULL,
  min_maf = NULL,
  min_call_rate = NULL
)
}
\arguments{
\item{vcf}{VCF object}

\item{min_per_genotype}{Minimum samples per genotype group}

\item{min_maf}{Minimum minor allele frequency}

\item{min_call_rate}{Minimum call rate}
}
\value{
Filtered VCF object
}
\description{
Filter VCF with multiple criteria
}
