% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_banovich.R
\name{get_banovich_SE}
\alias{get_banovich_SE}
\title{operate with BiocFileCache to retrieve SummarizedExperiment banovich 450k data}
\usage{
get_banovich_SE(ca = BiocFileCache::BiocFileCache())
}
\arguments{
\item{ca}{BiocFileCache instance}
}
\description{
operate with BiocFileCache to retrieve SummarizedExperiment banovich 450k data
}
\examples{
bano = get_banovich_SE()
bano
}
