% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_tensorqtl.R
\name{get_tensor_example_path}
\alias{get_tensor_example_path}
\title{operate with BiocFileCache to retrieve zip file of tensorQTL output}
\usage{
get_tensor_example_path(ca = BiocFileCache::BiocFileCache())
}
\arguments{
\item{ca}{BiocFileCache instance}
}
\description{
operate with BiocFileCache to retrieve zip file of tensorQTL output
}
\examples{
get_tensor_example_path()
}
