% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{getCalls}
\alias{getCalls}
\title{getter for genotype calls}
\usage{
getCalls(xse)
}
\arguments{
\item{xse}{instance of XqtlExperiment}
}
\description{
getter for genotype calls
}
\examples{
data(mageSE_19)
vp = system.file("vcf", "chr19_50k.vcf.gz", package="biocXqtl")
mp = minorAlleleCounts(vp)
xmage19 = XqtlExperiment(mageSE_19, mp)
getCalls(xmage19)[1:4,1:4]
}
