% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_utils.R
\name{getGenotypeSummary}
\alias{getGenotypeSummary}
\title{Get genotype summary statistics}
\usage{
getGenotypeSummary(vcf)
}
\arguments{
\item{vcf}{A VCF object}
}
\value{
DataFrame with genotype statistics per variant
}
\description{
Get genotype summary statistics
}
