% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geuv19}
\alias{geuv19}
\title{a RangedSummarizedExperiment with data from GeuvadisTranscriptExpr}
\format{
RangedSummarizedExperiment
}
\usage{
data(geuv19)
}
\description{
a RangedSummarizedExperiment with data from GeuvadisTranscriptExpr
}
\note{
Minor allele counts are in colData.  Some entries are -1.  This
is not explained in GeuvadisTranscriptExpr.
}
\examples{
data(geuv19)
rowRanges(geuv19[1:3,])
}
\keyword{datasets}
