% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geuv19_samples}
\alias{geuv19_samples}
\title{data frame with sample information for geuv19}
\format{
data.frame
}
\usage{
geuv19_samples
}
\description{
data frame with sample information for geuv19
}
\keyword{datasets}
