% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geuv19xse}
\alias{geuv19xse}
\title{a RangedSummarizedExperiment with data from GeuvadisTranscriptExpr, in XqtlExperiment format}
\format{
RangedSummarizedExperiment
}
\usage{
data(geuv19xse)
}
\description{
a RangedSummarizedExperiment with data from GeuvadisTranscriptExpr, in XqtlExperiment format
}
\note{
Minor allele counts are in calls slot.  Some entries are -1.  This
is not explained in GeuvadisTranscriptExpr.
}
\examples{
data(geuv19xse)
rowRanges(geuv19xse[1:3,])
}
\keyword{datasets}
