% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INT.R
\name{INT}
\alias{INT}
\title{inverse normal transformation}
\usage{
INT(x)
}
\arguments{
\item{x}{matrix for which rows are to be transformed}
}
\description{
inverse normal transformation
}
\examples{
m = matrix(rbeta(2000,2,1),4,500)
par(mfrow=c(2,2))
mt = INT(m)
hist(m[1,]); hist(mt[1,])
hist(m[2,]); hist(mt[2,])
}
