% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_xq_gwas.R
\name{join_xqtr_gwas}
\alias{join_xqtr_gwas}
\title{combine an output of tq_xexp_trans with a GWAS catalog-like resource}
\usage{
join_xqtr_gwas(
  transtab,
  gwascatdf,
  verbose = TRUE,
  phtag = "symbol",
  linkby = "rsid"
)
}
\arguments{
\item{transtab}{output of tq_xexp_trans}

\item{gwascatdf}{data.frame with CHR_POS, CHR_ID, SNPS (if using linkby='rsid')}

\item{verbose}{logical(1)}

\item{phtag}{character(1) name of variable in transtab regarded as molecular phenotype,
typically 'symbol' (default)}

\item{linkby}{character(1), "rsid" or "pos"}
}
\description{
combine an output of tq_xexp_trans with a GWAS catalog-like resource
}
