% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{simple function for MAF calculation}
\usage{
maf(snpmat)
}
\arguments{
\item{snpmat}{an entity that answers `apply(...,2,f)` with columns corresponding to SNPs,
rows corresponding to minor allele count for individuals}
}
\description{
simple function for MAF calculation
}
\examples{
data("geuv19", package="biocXqtl")
m = maf(colData(geuv19))
summary(m)
}
