% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorAlleleCounts.R
\name{minorAlleleCounts}
\alias{minorAlleleCounts}
\title{produce data.frame with minor allele counts in a region, from VCF}
\usage{
minorAlleleCounts(
  vcfpath,
  region = GRanges(c("19:1-100000", "19:100001-200000")),
  genome = "GRCh37"
)
}
\arguments{
\item{vcfpath}{character path to VCF}

\item{region}{a GRanges instance}

\item{genome}{character(1)}
}
\value{
a GRanges with mcols giving the count of minor alleles for each sample at each locus
}
\description{
produce data.frame with minor allele counts in a region, from VCF
}
\examples{
vp = system.file("vcf", "chr19_50k.vcf.gz", package="biocXqtl")
mp = minorAlleleCounts(vp)
mp[,1:5]
dim(mcols(mp))
}
