% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{show,XqtlExperiment-method}
\alias{show,XqtlExperiment-method}
\title{present concise view of XqtlExperiment}
\usage{
\S4method{show}{XqtlExperiment}(object)
}
\description{
present concise view of XqtlExperiment
}
