% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tens445}
\alias{tens445}
\title{SummarizedExperiment with example data from tensorQTL, with 445 individuals
from the GEUVADIS project}
\format{
SummarizedExperiment
}
\usage{
data(tens445)
}
\description{
SummarizedExperiment with example data from tensorQTL, with 445 individuals
from the GEUVADIS project
}
\note{
Associated genotype data is available in PlinkArray with `example_PlinkArray()`.
}
\keyword{datasets}
