% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiles38}
\alias{tiles38}
\title{a GRanges tiling GRCh38}
\format{
GRanges
}
\usage{
data(tiles38)
}
\description{
a GRanges tiling GRCh38
}
\note{
Created with `GenomicRanges::tileGenome(BSgenome.Hsapiens.UCSC.hg38)`
}
\keyword{datasets}
