% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorqtl_support.R
\name{tq_xexp_cis}
\alias{tq_xexp_cis}
\title{tq_xexp_cis runs tensorqtl cis.map_nominal on components from an XqtlExperiment instance}
\usage{
tq_xexp_cis(
  xexp,
  window = 1000000L,
  maf_threshold = 0,
  prefix = "tqrun",
  write_parquet_path = tempdir()
)
}
\arguments{
\item{xexp}{XqtlExperiment instance}

\item{window}{integer(1) region around start to search}

\item{maf_threshold}{numeric(1) defaults to 0.}

\item{prefix}{character(1) used as stem for output, defaults to 'tqrun'}

\item{write_parquet_path}{path to folder used for output, defaults to `tempdir()`}
}
\description{
tq_xexp_cis runs tensorqtl cis.map_nominal on components from an XqtlExperiment instance
}
\examples{
example(XqtlExperiment) # makes nn
td = tempdir()
sexn = ifelse(colData(nn)$sex == "XY", 0., 1.)
colData(nn) = NULL
nn$sex = sexn
tq_xexp_cis(nn, prefix="example_tq", write_parquet_path=td)
dir(td, full=TRUE)
}
