% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorqtl_support.R
\name{tq_xexp_trans}
\alias{tq_xexp_trans}
\title{tq_xexp_trans runs tensorqtl map_trans on components from an XqtlExperiment instance}
\usage{
tq_xexp_trans(
  xexp,
  pval_threshold = 1e-05,
  maf_threshold = 0.05,
  batch_size = 20000L
)
}
\arguments{
\item{xexp}{XqtlExperiment instance}

\item{pval_threshold}{numeric(1) defaults to 1e-05,}

\item{maf_threshold}{numeric(1) defaults to 0.05}

\item{batch_size}{numeric(1) defaults to 20000}
}
\description{
tq_xexp_trans runs tensorqtl map_trans on components from an XqtlExperiment instance
}
\examples{
example(XqtlExperiment) # makes nn
td = tempdir()
sexn = ifelse(colData(nn)$sex == "XY", 0., 1.)
colData(nn) = NULL
nn$sex = sexn
lk = tq_xexp_trans(nn, maf_threshold=.01, pval_threshold=1e-4)
head(lk)
}
