% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{XqtlExperiment-class}
\alias{XqtlExperiment-class}
\title{extend RangedSummarizedExperiment to include genotype calls in a GRanges}
\description{
extend RangedSummarizedExperiment to include genotype calls in a GRanges
}
\note{
We use RangedSummarizedExperiment to ensure we can identify genomic
distance between molecular features and variants
}
\examples{
data(mageSE_19)
vp = system.file("vcf", "chr19_50k.vcf.gz", package="biocXqtl")
mp = minorAlleleCounts(vp)
mp[,1:5]
mp = mp[, colnames(mageSE_19)]
new("XqtlExperiment", mageSE_19, calls=mp, NAMES=NULL)
}
