% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{XqtlExperiment}
\alias{XqtlExperiment}
\title{XqtlExperiment constructor}
\usage{
XqtlExperiment(se, calls)
}
\arguments{
\item{se}{SummarizedExperiment instance}

\item{calls}{GRanges instance}
}
\description{
XqtlExperiment constructor
}
\examples{
data(mageSE_19)
vp = system.file("vcf", "chr19_50k.vcf.gz", package="biocXqtl")
mp = minorAlleleCounts(vp)
nn = XqtlExperiment(mageSE_19, mp)
nn
}
