% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getzs.R
\name{getzs}
\alias{getzs}
\title{create a matrix of Z statistics for an additive effect of rare allele count in calls on molecular
phenotype in molec}
\usage{
getzs(molec, calls)
}
\arguments{
\item{molec}{numeric matrix of molecular phenotype values, columns are samples}

\item{calls}{numeric matrix of rare allele counts, rows are samples, columns are loci}
}
\value{
A matrix with one row per
}
\description{
create a matrix of Z statistics for an additive effect of rare allele count in calls on molecular
phenotype in molec
}
\note{
For a SNP with MAF 0, NA is returned.
}
\examples{
data(geuv19)
mol = assay(geuv19)
calls = data.matrix(as.data.frame(colData(geuv19)))
m = maf(calls)
allz = getzs(mol[1:100,], calls[,m>.3])
summary(as.numeric(allz))
}
