% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{maf}
\alias{maf}
\title{compute putative minor allele frequency for XqtlExperiment}
\usage{
maf(xse)
}
\arguments{
\item{xse}{XqtlExperiment instance}
}
\description{
compute putative minor allele frequency for XqtlExperiment
}
\examples{
example(XqtlExperiment)  # makes nn
summary(maf(nn))
}
