% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz_stats}
\alias{viz_stats}
\title{produce a plotly display of statistics of xQTL association}
\usage{
viz_stats(
  se,
  jitter_fac = 500,
  ptcolor = "blue",
  midchop = 2,
  xlabel = "SNP addr",
  ylabel = "xQTL association Z"
)
}
\arguments{
\item{se}{SummarizedExperiment post "bind_Zs"}

\item{jitter_fac}{plot employs jittering of SNP address by this factor}

\item{ptcolor}{point color}

\item{midchop}{stats with absolute value less than this value are omitted from display}

\item{xlabel}{axis label}

\item{ylabel}{axis label}
}
\description{
produce a plotly display of statistics of xQTL association
}
\examples{
data(geuv19)
sds = MatrixGenerics::rowSds(assay(geuv19), na.rm=TRUE)
qq = quantile(sds, .8)
ok = which(sds > qq)
lk = geuv19[ok,]
mafs = maf(colData(lk)) # only snps here
mins = apply(data.matrix(as.data.frame(colData(lk))), 2, min, na.rm=TRUE) # some -1 values
colData(lk) = colData(lk)[,which(mafs>.25 & mins > -1)]
lk <- bind_Zs(lk, colselector = function(se) colnames(colData(se)))
viz_stats(lk)
}
