% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zs4manyYs.R
\name{zs4manyYs}
\alias{zs4manyYs}
\title{use procedure tailored to multiple responses for a fixed design matrix}
\usage{
zs4manyYs(se, seq4mol = 0.5, minmaf = 0.3, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{se}{RangedSummarizedExperiment}

\item{seq4mol}{quantile of rowSds(assay(se)) at which to filter away features with low variability over samples}

\item{minmaf}{lower bound on MAF for genotype calls}

\item{BPPARAM}{instance of BiocParallelParam to control parallel execution over genotype calls}
}
\value{
matrix with rows corresponding to molecular features and columns corresponding to genotypes
}
\description{
use procedure tailored to multiple responses for a fixed design matrix
}
\examples{
data(geuv19)
m = maf(colData(geuv19)) # only genotypes
sds = MatrixGenerics::rowSds(assay(geuv19))
print(quantile(m)) # note negatives!
print(quantile(sds))
selz = zs4manyYs(geuv19, minmaf = median(m))
print(dim(selz))
print(selz[1:5,1:5])
}
