% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zs4manyYs.R
\name{zs4manyYs}
\alias{zs4manyYs}
\title{use procedure tailored to multiple responses for a fixed design matrix}
\usage{
zs4manyYs(se, omit.hi.maf = FALSE, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{se}{RangedSummarizedExperiment}

\item{omit.hi.maf}{logical(1) if TRUE, variants for which MAF exceeds 0.5 are omitted}

\item{BPPARAM}{instance of BiocParallelParam to control parallel execution over genotype calls}
}
\value{
matrix with rows corresponding to molecular features and columns corresponding to genotypes
}
\description{
use procedure tailored to multiple responses for a fixed design matrix
}
\examples{
data(geuv19xse)
sds = MatrixGenerics::rowSds(assay(geuv19xse))
print(quantile(sds))
BiocParallel::register(BiocParallel::SerialParam())
mafs = maf(geuv19xse)
mins = apply(data.matrix(mcols(getCalls(geuv19xse))), 1, min, na.rm=TRUE) # some -1 values
print(quantile(mins))
lk = filterCalls(geuv19xse, which(mafs>.25 & mins > -1))
lk = lk[which(sds>median(sds, na.rm=TRUE)),]
chk1 = zs4manyYs(lk)
data(geuv19_samples)
namedSex = geuv19_samples$Sex
names(namedSex) = geuv19_samples[["Sample name"]]
lk$Sex = namedSex[colnames(lk)]
table(lk$Sex)
chk2 <- zs4manyYs(lk) # use covariate
plot(as.numeric(chk2), as.numeric(chk1))
}
